allow for printing to console using [PrintStrategy]ies, deprecate printToConsole
