+++ pubspec.yaml
@@ -12,4 +12,4 @@ dependencies:
   logging: ^1.0.2
 
 dev_dependencies:
-  leancode_lint: 1.0.1+1
+  leancode_lint: ^1.0.2
